refactor(frontend): renamed scripts/ folder to vite/ since all files relate to vite
