+++ frontend/vite/vite-plugin-flatten-html.ts
