+++ frontend/vite/vite-plugin-route-rewriter.ts
